Logic cleanup in Buiding class
Fixes an initialization problem where having "items=[]" in the initializer constructor would cause that list to be reused across instances of the object instead of per object. To work around this the function will default the value to None then initialize it inside the method to make the list independent per instance.

Server sort functions called the wrong sort function in the parent class and were cleaned up.