Initial GUI Work
Adds Gui.py which offers a graphical driver for interacting with the inventory. This GUI offers:

- Functions to handle switching views
- A themed interface
- A view for the initial load-in page
- A view for viewing the loaded campus, with a grid based layout
- A modal warning for currently unimplemented functions

Much of the functionality in the GUI interface is still not implemented. "TODO" comments and modal popup warnings signal functionality that is stubbed. Additional views still need to be created as well.


On implementing this several changes were needed to other files:

- FileManager.py was reworked to allow taking a full file path instead of just a file name. The assumption the files will always be in the "Campus" directory was removed.
- test.py was updated accordingly
- Building.py and Campus.py had minor updates